refactor(agents): extract create_pxi_agent helper
Why: keep /chat-v2 router focused on request handling and centralize
PXI agent construction (deps, output, prompts, toolsets, history
processors) in a dedicated module.
