+++ src/phoenix/server/api/routers/chat_v2.py
@@ -6,9 +6,7 @@ from fastapi import APIRouter, Depends, HTTPException, Query
 from openinference.instrumentation import using_session
 from pydantic import BaseModel, Field
 from pydantic.types import Discriminator
-from pydantic_ai import Agent, AgentRunResult, DeferredToolRequests, RunContext
-from pydantic_ai.agent import InstrumentationSettings
-from pydantic_ai.messages import ModelMessage
+from pydantic_ai import AgentRunResult
 from pydantic_ai.ui.vercel_ai import VercelAIAdapter
 from pydantic_ai.ui.vercel_ai.request_types import (
     RegenerateMessage,
@@ -26,23 +24,15 @@ from phoenix.config import (
 from phoenix.server.agents.capabilities import AgentCapabilities
 from phoenix.server.agents.chat_params import ChatSearchParamsModel
 from phoenix.server.agents.chat_v2.dependencies import ChatDependencies
-from phoenix.server.agents.chat_v2.toolsets import build_chat_v2_toolsets
+from phoenix.server.agents.chat_v2.pxi_agent import ChatOutput, create_pxi_agent
 from phoenix.server.agents.context import (
     ChatContext,
-    build_phoenix_context_user_message_content,
-    insert_context_user_message,
     resolve_contexts,
 )
 from phoenix.server.agents.exceptions import AgentError
 from phoenix.server.agents.instrumentation import get_tracer_provider
 from phoenix.server.agents.model_factory import build_chat_model
-from phoenix.server.agents.prompts import (
-    AGENT_STATIC_SYSTEM_PROMPT,
-    build_agent_dynamic_system_prompt,
-)
 from phoenix.server.bearer_auth import is_authenticated
-
-ChatOutput = str | DeferredToolRequests
 
 
 class _ChatMessageMixin(BaseModel):
@@ -68,28 +58,6 @@ _RequestData = Annotated[
 
 
 logger = logging.getLogger(__name__)
-
-
-def _build_dynamic_instructions(ctx: RunContext[ChatDependencies]) -> str | None:
-    """Render request-specific PXI instructions from the run's dependencies."""
-    return build_agent_dynamic_system_prompt(capabilities=ctx.deps.capabilities)
-
-
-def _inject_ui_context(
-    ctx: RunContext[ChatDependencies],
-    messages: list[ModelMessage],
-) -> list[ModelMessage]:
-    """Append the per-turn Phoenix UI context as a trailing user message.
-
-    Running as a history processor (rather than mutating the request body)
-    keeps the static prefix — system prompt + prior conversation history —
-    byte-identical across turns, which is the prerequisite for any
-    provider-side prompt cache to take effect.
-    """
-    return insert_context_user_message(
-        messages,
-        build_phoenix_context_user_message_content(ctx.deps.contexts),
-    )
 
 
 def _log_run_complete(result: AgentRunResult[Any]) -> None:
@@ -132,20 +100,7 @@ def create_chat_v2_router(authentication_enabled: bool) -> APIRouter:
             collector_api_key=get_env_phoenix_agents_collector_api_key(),
             project_name=get_env_phoenix_agents_assistant_project_name(),
         )
-        instrument = (
-            InstrumentationSettings(tracer_provider=tracer_provider)
-            if tracer_provider is not None
-            else None
-        )
-        agent: Agent[ChatDependencies, ChatOutput] = Agent(
-            model,
-            deps_type=ChatDependencies,
-            output_type=[str, DeferredToolRequests],
-            instructions=[AGENT_STATIC_SYSTEM_PROMPT, _build_dynamic_instructions],
-            toolsets=[lambda ctx: build_chat_v2_toolsets(ctx.deps)],
-            history_processors=[_inject_ui_context],
-            instrument=instrument,
-        )
+        agent = create_pxi_agent(model, tracer_provider=tracer_provider)
         adapter: VercelAIAdapter[ChatDependencies, ChatOutput] = VercelAIAdapter(
             agent=agent,
             run_input=body,
